\name{interfaceProxy}
\alias{interfaceProxy}
\title{Create the Java object corresponding to the interface proxy.}
\usage{
  interfaceProxy(interface, implementation)
}
\arguments{
  \item{interface}{the interface to proxy}

  \item{implementation}{the implementing proxy}
}
\value{
  Java interface-proxy
}
\description{
  Create the Java object corresponding to the interface
  proxy.
}

