\name{setJavaRefClass}
\alias{setJavaRefClass}
\title{Create a refClass-based proxy for Java classes.}
\usage{
  setJavaRefClass(className, where =
  topenv(parent.frame()))
}
\arguments{
  \item{className}{the class to be proxied}

  \item{where}{environment to define the refClass in}
}
\value{
  a refClass-based Java-class-proxy.
}
\description{
  Create a refClass-based proxy for Java classes.
}

