\name{interfaceProxy}
\alias{interfaceProxy}
\title{Instantiate a Java-based interface-proxy for the...}
\usage{interfaceProxy(interface, implementation)}
\description{Instantiate a Java-based \code{interface}-proxy for the
\code{implementation}.}
\value{a suitable \code{RInterfaceProxy}}
\arguments{\item{interface}{the name of the to-be-implemented Java interface}
\item{implementation}{an instance of the implementing \code{refClass}}}
